Ensure keyboard is closed when activity ends.
